pyproject.toml
[tool.poetry]
name = "credit-card-management-backend"
version = "0.1.0"
description = ""
authors = ["PedroHMNobrega <pedrohmnobrega@gmail.com>"]
readme = "README.md"
packages = [{include = "setup"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
